Projectiles now damage the player/enemies. Use lightning strike by pressing Q.

+++ Assets/Scripts/Enemies/Enemy.cs
@@ -6,13 +6,13 @@ public class Enemy : MonoBehaviour {
 
 	[System.Serializable]
     public class EnemyStats {
-        public int health = 100;
+        public float health = 100f;
         public float contactDamage = 20f;
     }
 
     public EnemyStats stats = new EnemyStats();
 
-    public void DamageEnemy(int damage) {
+    public void DamageEnemy(float damage) {
         stats.health -= damage;
         if (stats.health <= 0) {
             GameMaster.KillEnemy(this);
